refactor: 채팅, 내정보, 설정 페이지 추가 및 구조 개선
