+++ app/src/main/java/com/apptive/japkor/ui/home/mypage/MypageScreen.kt
@@ -1,0 +1,24 @@
+package com.apptive.japkor.ui.home.mypage
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import com.apptive.japkor.ui.components.CustomText
+import com.apptive.japkor.ui.components.CustomTextType
+import com.apptive.japkor.ui.theme.CustomColor
+
+@Composable
+fun MypageScreen(modifier: Modifier = Modifier) {
+    Box(
+        modifier = modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        CustomText(
+            text = "내정보 페이지입니다.",
+            type = CustomTextType.body,
+            color = CustomColor.gray400
+        )
+    }
+}
